Fix type of np: it should be size_t

+++ src/xc.h
@@ -294,7 +294,7 @@ extern const int xc_input_variables_family_key[]; /* mapping input variable -> f
 extern const int xc_input_variables_flags_key[];  /* mapping input variable -> flags */
 
 const xc_input_variables_dimensions *input_variables_dimensions_get(int nspin);
-xc_input_variables *xc_input_variables_allocate(double np, int family, int flags, int nspin);
+xc_input_variables *xc_input_variables_allocate(size_t np, int family, int flags, int nspin);
 int xc_input_variables_sanity_check(const xc_input_variables *out, int family, int flags);
 void xc_input_variables_initialize(xc_input_variables *out);
 void xc_input_variables_deallocate(xc_input_variables *out);
@@ -307,7 +307,7 @@ extern const int xc_output_variables_flags_key[];  /* mapping output variable ->
 extern const int xc_output_variables_order_key[];  /* mapping output variable -> order of derivative */
 
 const xc_output_variables_dimensions *output_variables_dimensions_get(int nspin);
-xc_output_variables *xc_output_variables_allocate(double np, const int *orders, int family, int flags, int nspin);
+xc_output_variables *xc_output_variables_allocate(size_t np, const int *orders, int family, int flags, int nspin);
 int xc_output_variables_sanity_check(const xc_output_variables *out, const int *orders, int family, int flags);
 void xc_output_variables_initialize(xc_output_variables *out);
 void xc_output_variables_deallocate(xc_output_variables *out);
